+++ argilla/src/argilla/settings/_resource.py
@@ -14,15 +14,17 @@
 
 import json
 import os
+import re
 from functools import cached_property
 from pathlib import Path
-from typing import List, Optional, TYPE_CHECKING, Dict, Union, Iterator, Sequence
+from typing import List, Optional, TYPE_CHECKING, Dict, Union, Iterator, Sequence, Literal
 from uuid import UUID
 
 from argilla._exceptions import SettingsError, ArgillaAPIError, ArgillaSerializeError
 from argilla._models._dataset import DatasetModel
 from argilla._resource import Resource
 from argilla.settings._field import Field, _field_from_dict, _field_from_model
+from argilla.settings._io import build_settings_from_repo_id
 from argilla.settings._metadata import MetadataType, MetadataField
 from argilla.settings._question import QuestionType, question_from_model, question_from_dict
 from argilla.settings._task_distribution import TaskDistribution
@@ -258,6 +260,25 @@ class Settings(DefaultSettingsMixin, Resource):
             settings_dict = json.load(file)
             return cls._from_dict(settings_dict)
 
+    @classmethod
+    def from_hub(
+        cls,
+        repo_id: str,
+        subset: Optional[str] = None,
+        feature_mapping: Optional[Dict[str, Literal["question", "field", "metadata"]]] = None,
+        **kwargs,
+    ) -> "Settings":
+        """Load the settings from the Hub
+
+        Parameters:
+            repo_id (str): The ID of the repository to load the settings from on the Hub.
+            subset (Optional[str]): The subset of the repository to load the settings from.
+            feature_mapping (Dict[str, Literal["question", "field", "metadata"]]): A dictionary that maps incoming column names to Argilla attributes.
+        """
+
+        settings = build_settings_from_repo_id(repo_id=repo_id, feature_mapping=feature_mapping, subset=subset)
+        return settings
+
     def __eq__(self, other: "Settings") -> bool:
         return self.serialize() == other.serialize()  # TODO: Create proper __eq__ methods for fields and questions
 
@@ -381,7 +402,7 @@ class Settings(DefaultSettingsMixin, Resource):
                 dataset_properties_by_name[property.name] = property
 
     @classmethod
-    def _validate_mapping(cls, mapping: Dict[str, Union[str, Sequence[str]]]) -> None:
+    def _validate_mapping(cls, mapping: Dict[str, Union[str, Sequence[str]]]) -> dict:
         validate_mapping = {}
         for key, value in mapping.items():
             if isinstance(value, str):
@@ -390,7 +411,17 @@ class Settings(DefaultSettingsMixin, Resource):
                 validate_mapping[key] = tuple(value)
             else:
                 raise SettingsError(f"Invalid mapping value for key {key!r}: {value}")
+
         return validate_mapping
+
+    @classmethod
+    def _sanitize_settings_name(cls, name: str) -> str:
+        """Sanitize the name for the settings"""
+
+        for char in [" ", ":", ".", "&", "?", "!"]:
+            name = name.replace(char, "_")
+
+        return name.lower()
 
     def __process_guidelines(self, guidelines):
         if guidelines is None:
@@ -404,6 +435,11 @@ class Settings(DefaultSettingsMixin, Resource):
                 return file.read()
 
         return guidelines
+
+    @classmethod
+    def _is_valid_name(cls, name: str) -> bool:
+        """Check if the name is valid"""
+        return bool(re.match(r"^(?=.*[a-z0-9])[a-z0-9_-]+$", name))
 
 
 Property = Union[Field, VectorField, MetadataType, QuestionType]